Tests moved to a separate project
